+++ src/dialects/mod.rs
@@ -1378,6 +1378,7 @@ fn transform_quotes_in_table_source(source: &mut TableSource, target: Dialect) {
         }
         TableSource::Subquery { .. } => {}
         TableSource::TableFunction { .. } => {}
+        TableSource::Raw { .. } => {}
         TableSource::Values { .. } => {}
         TableSource::Lateral { source } => transform_quotes_in_table_source(source, target),
         TableSource::Pivot { source, .. } | TableSource::Unpivot { source, .. } => {